refactor: fold session-product batch hydration into refresh and rebuild paths
Inline the remaining batch helper layer once refresh/rebuild call sites have stabilized around the consolidated product modules.
